Assignment 5B: Calculator UI
Added float functionality to all calculator functions in 5B. For extra cinnamon and spice, you must select an operation type before you are allowed to calculate, and the calculator yells at you for empty inputs. Also, the output boxes do not appear until you calculate for the first time!

Also fixed a bug where Subtract was turning the initial value given into a negative. This reflects in both 4 and 5B.